requirements.txt
aiofiles==23.2.1
aiogram==3.3.0
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
colorama==0.4.6
contourpy==1.2.0
cycler==0.12.1
environs==10.3.0
filelock==3.13.1
flake8==7.0.0
fonttools==4.47.2
frozenlist==1.4.1
fsspec==2023.12.2
idna==3.6
iniconfig==2.0.0
Jinja2==3.1.3
kiwisolver==1.4.5
magic-filter==1.0.12
MarkupSafe==2.1.4
marshmallow==3.20.2
matplotlib==3.8.2
mccabe==0.7.0
mpmath==1.3.0
multidict==6.0.5
networkx==3.2.1
numpy==1.26.3
packaging==23.2
pillow==10.2.0
pluggy==1.4.0
pycodestyle==2.11.1
pydantic==2.5.3
pydantic_core==2.14.6
pyflakes==3.2.0
pyparsing==3.1.1
pytest==8.0.0
python-dateutil==2.8.2
python-dotenv==1.0.1
requests==2.31.0
six==1.16.0
sympy==1.12
torch==2.2.0
torchvision==0.17.0
tqdm==4.66.1
typing_extensions==4.9.0
urllib3==2.2.0
yarl==1.9.4
